requirements.txt
enhanced-dpy
aiohttp
jishaku
discord-py-slash-command
